SPARTA: Remove the custom_site_config hackery
At some point we mostly-implemented a feature for the /internal panel
that let you preview a theme without applying it globally - the idea
being that it'd be useful for us when creating a theme. We don't
actually use that functionality, however - it's much more common to
develop a theme in a local installation, including adding new image
assets for each customer, and then copying the config when we're
satisfied with it.

It's also the source of a nasty chicken/egg problem with the middleware
- we need the session loaded when we get the current domain because we
were checking for the custom_site_id, and we need the current domain
loaded in order to determine HTTP/HTTPS enforcement, and we need the
HTTP/HTTPS enforcement loaded before the session store so it can rewrite
the cookies. Killing the custom_session_id broke the dependency loop.

QA: Shouldn't affect anything you test. On /internal, you'll no longer
be able to preview the non-active theme.

Reviewed-by: Aiden Scandella <aiden.scandella@socrata.com>