BUG: 4397 Frame busting for cross-domain iframes in IE, Firefox
Somewhere down the line our ability to bust out of frames in IE and
Firefox broke, potentially exposing us to a vulnerability where a
malicious site could embed the socrata login page as a full-page IFRAME,
but capture the username/password from the embedded IFRAME. I don't
believe this is actually possible in practice - while the site wouldn't
bust out of the IFRAME, it didn't do so specifically because the browser
was blocking attempts at malicious javascript - the same code that would
block a malicious attack also blocked the ability to frame bust.

The new solution is twofold:
1. Wrap the frame-busting code in a try/catch block, so that an error in
IE/FF is caught and we just assume we're in a frame and attempt to bust
it. This works for older IE browsers that don't yet support option #2.
2. Send a new HTTP header (X-Frame-Options: SAMEORIGIN) on almost all
pages to prevent the ability to embed our site in another site at a
higher level. In modern (IE8+, FF 3.6.9+) browsers, this is a much more
reliable way to accomplish the same end-goal, even though the user
experience is entirely different.

QA: I attached a simple HTML page you can use as a test case in bug
4397. You'll want to regression check that embeds (widgets, browse
embed, etc.) continue to work, and you'll want to verify this in as many
browsers as possible. In general, with the test case HTML, you should
see one of two things: the page loads briefly in the frame then the
entire page reloads with the same content, or the frame doesn't load at
all - depending on the browser you'll either see a white box or some
browser warning that the content can't be embedded, often with a link to
open it in a new window.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>