FEATURE: Open a sidebar pane once by default.
QA: Go to a dataset, and add #pane=about to the end
of the URL. Make sure that the about sidebar pops up
when the page loads. After that, it should behave as
before (if you have a filter, it will go back to the
filter pane when you close a bar, etc); it's just the
moment you load the page that should change.

To-be-
Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>