BUG: Fix issue with out-dated dataset metadata being returned and used If a new row request with updated metadata started while an older one was still in process, the old one would return and be used, causing a state change back to a previous state. Now keep track of the latest meta request fired off, and ignore and old ones that come back
QA:
In a map, toggle on and off several of the render types quickly (while
previous things are still loading). Previously, it would be easy to
get things stuck in a state where render types kept randomly opening
and closing after you stopped clicking. That should no longer happen
(but it might still take a bit of time for your browser to catch up).

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>