FEATURE: Add columnId listing to API sidebar.
QA: Under Export => API on the sidebar, there will
be a new section at the bottom listing out the column
name and ID for each dataset. Make sure it works.

More importantly, this change touched the styles for
the about sidebar and about page, so make sure that
those didn't break as well.

Reviewed-by: Aiden Scandella <aiden.scandella@socrata.com>