BUG: Authentication failure using inline login to post a comment on a dataset
Our Rack-SSL-Enforcer middleware conflates the concept of what sites
must be redirected to SSL endpoints and what endpoints must only return
secure (HTTPS-only) cookies. Our normal login flow POSTs to
/user_sessions which doesn't require HTTPS (but probably should), and
therefore allows the _blist_session_id cookie to be used on a non-HTTPS
page. However, inline login does a GET to /login.json which was
SSL-enforced.

In the future, we should make an attempt to make ALL requests to modify
content on our service use HTTPS, regardless of whether the domain is
HTTPS-enforced. This would get us closer to being able to close the hole
of having a session cookie available over unencrypted connections.

QA: This bug only exists on domains that do not have HTTPS-everywhere
enabled - in other words, don't test it on opendata.test-socrata.com. To
repro:

1. Log out
2. Visit a public dataset
3. Post a comment
4. You'll be asked for your credentials; login here and allow the post
to complete. (Before you'd get a browser popup asking for your username
and password, and if you cancel that a red error message indicating
authentication was required - but you'd still get the login/logout links
at the top to flip)

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>