BUG 4257: Location fields in import don't update state.
QA: The issue is that the underlying state doesn't get
updated when you tweak the column select dropdowns in
location imports, only when you change the main dropdowns
or type in text fields.
So, if the very last thing you do before hitting 'next'
is changing one of the location column source dropdowns
(say, address) to something else, make sure that that
last change is reflected.

Reviewed-by: Aiden Scandella <aiden.scandella@socrata.com>