FEATURE: Deleting comments for Wally.
QA: If you are the author of a comment, or you have
comment moderation privileges on the domain, you should
be able to delete comments. When a comment is deleted,
any comments nested beneath it should be deleted,
regardless of who authored them.

This feature should also work for nomination comments.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>