FEATURE 4913: Data context picker now does a server search for datasets * Context picker always sets the list of currently available datasets.   It then does a server search on name, and displays those results   together with the available items * Each set of items is under a header, and sorted appropriately * Available data returns immediately, server search is on a delay * Fix styles so awesomecomplete can go outside the sidebar * Add a search function to Dataset * Awesomecomplete changes: ** Make right alignment work correctly ** Add an option to ignore blank values, which makes items without a    value unselectable
QA: In the components editor, when picking a view/dataset for a
component, you should be able to type and get results (from the server)
for items that have that word in their name. Once you choose an item, it
should appear under 'Currently Used Data' in the list and show up much
more quickly (for example, when picking a view for a different
component).

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>