BUG 5340, z795: Minor fixes *  Fix snapshot sorting (would never get to publicationDate case) * Rich Render editor: Don't overflow:visible the renderArea for menus,   because this caused serious problems when the user had scrolled down.   Menus flip directions to stay in the container pretty well now, so it   isn't really needed anymore.
QA: Make sure bug is fixed.
Open the Rich Display editor for a dataset. Make sure you have a long
config that requries you to scroll to see the bottom items. Remove an
item near the bottom (when you are scrolled), and make sure the items
don't jump around, and you don't lose the scrollbar.

Reviewed-by: Aiden Scandella <aiden.scandella@socrata.com>