FEATURE: Configuration for series columns * Add section in chart config * Remove chart-create hack to preserve seriesColumns * In basePane, don't throw a validation error if a selectable section is disabled * Disable series col config if >500 rows
QA: When creating a chart (not donut, pie, or treemap), you should now
have an option to choose 'Data Series Group' columns. The values in this
column will be used to group the data into different series. For
example, the data might be:
Year,  State,  Population
2011,  WA,     54321
2010,  WA,     43210
2009,  WA,     32100
2011,  NY,     987654
2010,  NY,     876543
2009,  NY,     765432
Year would be the category column; Population would be the value column,
and State would be the data series group column, and you would end up
with a line for WA from 2009-2011, and a second line for NY for the same
years. If you don't use this column, charts should behave just as you
expect.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>