BUG: z1083 Short URLs can't strip the 'www' off anymore.
Our customers with a 'www.______.tld' domain have previously had their
"Short URL" for a dataset shortened to not use the "www" part in the
domain name. This *should* work, but often doesn't. If their redirection
server isn't setup right, it won't work, and if it's an HTTPS
connection, it won't work unless they have the bare domain setup for
SSL.

QA: If you have a test site that starts with "www." (I don't think you
do, unfortunately) you'd notice that the short URLs no longer remove the
www.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>