FEATURE: "manyTable" display type for Atlas datasets
* Added "manyTable" display type to render-type-manager.js. It is very
similar to the "table" type for now and actually only supports one table,
but this is my first stab at it. Eventually it will support choosing
which table out of the many child views you want to render.

QA:
Atlas datasets should now have a table view option that displays the
attributes for features in a table. The attributes should match what
pops up when you actually click on a feature on the map and get a popup.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>