FEATURE:-ish Show dataset/view listing in profile allways
QA: The dataset listing should now show in your profile regardless of whether or not you have datasets or views, and regardless of whether you can create new datasets or not.

Create a new user account that doesn't have permission to create datasets on a domain, and confirm that you see an empty dataset listing. And confirm that the listing shows properly for users that own datasets and views. That should probably be enough.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>