FEATURE: Only load calendar events for current view * Add flag check in baseViz to ignore change events * When the view is being changed on a viz, force a reset * In calendar, query for events within the current view and only render those * On initial load, view change, or when the query changes, auto-jump to the nearest event to now; unless there is an event within the view the user is looking at already * If auto-jumping, and no events are in the current view, then expand the query window a couple times to find the closest event without loading too many items (hopefully) * Aggressively mark times to ignore view changes in the calendar * Back to 500 row limit
QA: Large calendars should perform a bit better; in particular, if there
are over 500 events total, you should still get everything for the
current month you are looking at. It will load each time you switch
months, and only show the current month of data.
Test loading a calendar that doesn't have any events in the current
month, and make sure it initially auto-switches to show the nearest
event.
Make sure you can still switch months after that, and it displays a
message if there is no data for the current month.
Test filtering events when looking at a calendar. If you filter out
everything in the current month, it should auto-switch to the nearest
event. But if you change the filter, and there is still data in the
current month, the calendar should stay on the current month.

Reviewed-by: Aiden Scandella <aiden.scandella@socrata.com>