BUG z1231: Highcharts: only re-index data that is not data-grouped Data series grouping gives specific indices to each point to keep it together with the appropriate x value. Re-indexing it messes up that order, when it already has a valid index. Also fix null bars with sparse data.
QA:
Test with this data:
https://data.sunlightlabs.com/dataset/Super-Tuesday-Chart/whta-ftmh
Create a chart with State as the x-axis, Amount as the y-axis, and
'Super PAC' as the 'Group by' column under 'Data Series Grouping'. The
bars should correctly correspond to the state label on the x-axis.
Also test data series grouping on other charts, particularly in Chrome,
and make sure the data is not out-of-order on the x-axis.
Test null values and make sure the null bars appear.

Reviewed-by: Michael Chui <michael.chui@socrata.com>