FEATURE: Do row filtering on the client when possible * RowSet keeps track of many rows loaded to know if it is complete * RowSet optionally takes a parent * If parent is complete, then grab rows from the parent and filter those on the client * Rather complex logic to determine if one query can be derived from another (is a subset of, basically) * When creating a new row set, try to find one to derive from * Invalidating row sets should clear out all non-used ones * When translating a query, clean up extra levels to reduce it to the simplest form
QA: In a dataset where all the rows have been loaded, filtering to a
smaller set of rows should be fast (it will happen in the browser,
instead of going back to the server). Test large datasets where not all
the rows have been loaded, and make sure those still work. Also try
opening a saved filtered view and removing the filter to make sure it
correctly shows a larger set of rows. Sorting and grouping should also
work properly.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>