BUG: Clean-up/improve Canvas repositioning after changing size * Repeater used to attempt this; remove that code * More general solution that ties directly to window * Saves percentage position on scroll (hopefully only for user actions) * Restores that percentage position on a sizeRenderRefresh * Attempts to mask out scrolls caused by size changes
QA: When looking at an HSTP measure page, the links at the top should be
more reliable at showing you the chart you click on. There will be jumps
as the charts load, but once things are loaded, you should see the chart
you chose to jump to (it may not be perfectly centered, but should be
mostly visible).

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>