BUG: Fix render scroll scaling in d3 charts.
QA: This will only work in Chrome anyway (I still have to
fix a bug with the other browsers) but if you make a million-row
single-series (single value column) column chart it used to be
that if you scrolled far enough right the position the screen was
at didn't necessarily line up with the bars that were rendered.
It should be accurate now.

To-be-
Reviewed-by: jeff.scherpelz@socrata.com