NOOP: Make Y Axis Precision toggle between Auto and Slider.
QA: In charts side bar => Y Axis Precision, make sure the precision
setting works as you would expect. Before, it was just a slider
that defaults to 2. For nextgen=true, it should be a toggle between
Auto and a slider. Before, it would only kick the precision slider
in if the abbreviate checkbox was checked. Now, it'll respect each
setting separately.

To-be-
Reviewed-by: jeff.scherpelz@socrata.com