BUG: Cave and match wrong core server behavior until it changes: use header index rather than count on import.
* I don't think this is the right behavior, because using an index rather
than a count does not give the option of expressing "no headers." The core
server should consider "no headers" to be a possibility, eg if the first row
itself matches the type guess of the rest of the rows and at least some of
the type guesses are numeric.

QA: In almost all cases, the first row will automatically be considered a
header now.

To-be-
Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>