BUG: In rare cases, "do you want to save" dialog would pop when it shouldn't.
QA: I don't have an exact use case, but if you added 
for example ?pane=visualize# at the end of a dataset 
url, it would maybe pop the dialog when you try to press 
apply in the sidebar? It happened randomly to me and 
I don't really know the cause.

To-be-
Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>