FEATURE 6559: Visual styling for Canvas Designer * Corner icon to initiate edit * Main page inset * Links restyled to buttons in header * Added Settings button
QA: On a Canvas page you will now get an icon in the upper-left to
initiate Designer mode, instead of a black bar. On switching to
Designer, the main content of the page will look inset. The page name
and buttons will appear at the top of the page. Undo, Redo, and Settings
will have icons in addition to the text on the button. All buttons
except for Settings should still work.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>