FEATURE 6664: Use current filter for Download component * Option to disable the feature via disableInlineDownload * formInliner & downloadToFormCatcher no longer depend on blist.dataset or renderManager/table/etc * Add form-inliner to a separate JS package to dynamically load * Download comp is no longer fully rendered server-side b/c of the form inliner hookup
QA: Test the Download link on the HSMP measure page. It should still
work; but if you change filter conditions (including the top filters
like Year), only the data present in the filter should be in the
downloaded file.
Also on the normal dataset & widget pages, Download and Print should
work as normal.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>