BUG 6867: Enable data context picker for Text & Title * Fixed bug where searched datasets weren't showing up in context picker * Fixed bug where a non-required context picker would still validate as required * Reworked properties editor toggle on/off so that sidebar handles it   for both properties editor & properties palette. This makes it so   setting the data context updates the available properties immediately. * Update _propWrite so they properly handle changes to properties   (particularly dataContext) while in edit mode
QA: When editing a Title or Text component, you should now get a 'View
Data' option under 'Edit Component' in the sidebar. It will allow you to
search for a dataset, and once chosen, you should get a list of dataset
properties in the 'Available Properties' section of the sidebar that can
be dragged in to the component.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>