FEATURE 6660: Add custom display order to plain text columns * Similar to drop down list, but for any plain text column * _Not_ data sorting order (core server ignores it) * Used to control sort order for InlineFilter & chart series grouping   when an alphabetical sort is desired * Defaults to top values in order * basePane enhanced to accept an array of defaultValue for a repeater * Option on a selectable section to show it as long as there is data,   even if it matches the default
QA: In a plain text column, you should now have a 'Display Order'
section in column properties. This defaults to the top values in order
of prevalance (same order as 'Suggested Values' in the filter). You
should be able to change the items, and the changes should stick. It
shouldn't affect anything you can test, though.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>