BUG 6889: Re-enable adding components from sidebar in Designer * Real pane for components that goes in real sidebar * Make components default sidebar so they show by default * Improvements to sidebar ** show() shows default ** hide(true) forces a close, even if default exists ** won't re-show existing pane ** Don't show empty tip if no subtitle for pane * New module for future Designer features; currently components sidebar
QA: Canvas Designer should act as before by default. Adding the
'canvasX' module to the domain and enabling it as a feature will give
you access to upcoming features (before we release them to customers).
In this case, that is showing the Components sidebar, which is the
palette of components for dragging into the page. This sidebar should
show by default, be hidden when a component is selected, and re-show
when the component is deselected.  Exiting out of the Designer should
completely hide the sidebar.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>