FEATURE 7086: New 'API View' type * New displayType * New functions, values, strings, displays class, etc * In browse, use doc URL as main URL; add Manage button in table view * Dataset page is forced to grid, has no other view options * Disable saving any filter/search/sort changes * Many sidebars are hidden * Set new displayType when creating API view * Fix crash in API creater
QA: When you create an API via the API Foundry wizard, the resulting
view will be called an 'API View', and will have a different icon. They
should appear in the catalog and be filtered separately. By default the
link from the catalog will go to the Developer Docs; in the table view
of browse, expanding the section will show a Manage button that will
take you to the dataset page for the view. This will onlh be a grid
view, and have a lot of the standard tools disabled. It will still have
sharing & permissions management, along with editing the API. You will
be able to filter and look at the results, but not save a filtered view.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>