BUG 7207: Group bys suck; so don't try to handle them client-side
QA: Create a filtered view on a dataset with a couple of sorts and a
filter. Then group by one column, roll-up count by a second column, and
make sure neither of these two columns is used in the filter or sorts.
Save that grouped view. Then create a visualization on it, and make sure
the rows load.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>