BUG: in embedded HTML forms (not frames to our form views), setting the form input names to our column fieldName instead of the column id.
QA: using the embed pane in the grid view, create an form view, and then
an HTML form.  Save this HTML form into a separate web page.  In the
original dataset, edit the dataset and republish.  In your saved HTML
form, try to add some values and submit the form.  The republished
dataset should receive the values you submitted. This should fix
https://redmine.socrata.com/issues/7217

Reviewed-by: jeff.scherpelz@socrata.com