BUG: Handle malicious/incomplete referers without crashing
This fixes https://socrata.airbrake.io/groups/52942690

The referer HTTP header is sent by the client and as such needs to be
treated as untrusted user input. In some cases, the referer will exist,
but after parsing it with the URI library we won't have a host
component, for example:

  URI::parse('/foo').host # => nil

This makes us resilient to this particular error, which isn't
particularly common but the fix is easy.

QA: tamper with your HTTP request headers, loading a page that sets the
Referer: header to something without the hostname component, such as
'/foo'. You can accomplish this with either the Tamper Data extension
(Firefox-specific) or more generally using the Fiddler2 web debugging
proxy (Works with any Windows-based browser)

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>