BUG: Don't crash viewing profile pages on bad user input
This fixes https://socrata.airbrake.io/errors/53443179 - aka repro URL
https://data.cityofchicago.org/profile/Jamyia/is6y-5c5n?+Finance=&+Sustainable+Development=&limit&tags=std

I'm not really sure HOW we ended up at that URL (it's broken in more
ways than one) but we really should be validating user input to ensure
that it's int-like. Lucky for us, even nil.to_i == 0 in Ruby.

QA: Go to a profile page. Add "?limit" to the end of the URL. It
shouldn't crash.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>