FEATURE 7827: Management list for DataSlate pages * New section in admin * Lists pages & paths * Icons for published/unpublished, export, and private data * View/Download buttons for pages without params * Anything with params we don't know the proper fill-ins to view * Main links go directly to edit mode for everything * Edit mode is like debug mode (ignores errors & download behavior, and renders the page anyway); but doesn't show debug content, and immediately goes into edit mode * Removed random extra </div> on admin/users
QA: If you have rights to edit Slate pages, you should see a new section under Administration for DataSlate Pages. This will list all the pages, with icons for published/unpublished, export (download icon), and if a page is marked as private data. Clicking the main link of the page name should open the page and immediately go into DataSlate edit mode. For pages without parameters in the path, there will be a 'View Page' button on the right; or 'Download as CSV' for export pages.

To-be-
Reviewed-by: Clint Tseng <clint.tseng@socrata.com>