FEATURE 7837: Share component supports current page * currentPage property * Facebook, Twitter, email links for URL & name of current page * Hide subscribe * Email link is just mailto * Fixed bug with getting path from Canvas2::Util * Fixed bug with asset loading of same modal multiple times
QA: Check 'Linked Map & Chart' test page. There should be small buttons
to share at the very top of the page. It should include Twitter,
Facebook, and email; but not subscribe. Clicking the email link should
pop up a new email with a subject & body, but no address in the To
field.

To-be-
Reviewed-by: Clint Tseng <clint.tseng@socrata.com>