FEATURE 7826: Ability to create a new DataSlate page * Create button on DataSlate list (in admin) * Adapt configurator settings for create page * On POST, validate that path doesn't conflict, and if so redirect back with values to support editing * On create, redirect to new page in edit mode * Added option to modal help to configure style, in case you don't want it hidden initially * Methods in Page for checking for a dupe path, and creating a Page with reasonable defaults * Moved configuratorSettings styles into common file
QA: You should be able to go into the admin section for DataSlate Pages
and find a button to Create a new page. You can set the name and path
for a page, and it should give you an error if that path already exists.
If not, it should create the page and redirect you to it so that it
starts in edit mode.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>