BUG: Throw the proper exception when trying to load Users/ViewService methods and not logged in properly
This should fix http://support.socrata.com/tickets/2300

Note that this change requires a corresponding change in the core server
codebase as well, but they're independent and can be deployed in either
order - it just won't be fixed until both changes are released.

QA: Access
https://opendata.test-socrata.com/profile/anu/3nte-ch42?ownership=sharedToMe
when you are not logged in. It should redirect you to the login
page.

Access the same page while logged in as anyone other than Anu's account.
It should not display the datasets shared to you; I expect it'll
probably show a "Site Currently Unavailable" message and in this case
that's okay.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>