FEATURE: GovStat: Server communication * Goals are fetched from goal service for home page, goal page, Edit Goals * Got Backbone goals creating working * Added Save button * Still in-progress
QA: On govstat.test, you should be able to go to Edit Goals, create a
new goal (can only be in draft mode), hit Save, and the goal should
appear on reloading the page, and on loading the homepage. Updating a
goal may or may not work; putting it in a category doesn't work, and you
can't create a new category yet.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>