BUG: Crash in custom_content_controller displaying proper exception
This is a fix for https://socrata.airbrake.io/groups/55112536

The exception handling logic assumes that all exceptions that you can
catch here have an original_exception method. That's only true of
specific exception classes within the Rails framework, such as
ActionView::Template::Error. In this particular case for Airbrake, we
got a standard Ruby exception (NoMethodError) which doesn't respond to
original_exception, so now our exception handling code crashed.

QA: This was triggered at the URL
https://strategicplanning.aspe.hhs.gov/crosswalk-export.csv which means
that it was some particular custom crosswalk report. I don't actually
know how to reproduce this... and even if you could reproduce it, the
functional change here is that the site still crashes, but it crashes in
a prettier manner that shows the reason for crashing to the user.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>