BUG 9159, 9156, 9199: GovStat: Various fixes to goals editor * Make goal saving with metrics actually work * If a metric is not complete, pull it out and save in metadata * Properly determine completeness for Metric and Indicator * Also fix up parse/toJSON for above to work correctly * Add date_column on indicator for in model & UI * Hard-code some Indicator & Metric items that are not yet exposed * Reorganize editor to promote Start Date, and default it to today * Move Related Datasets & Additional Info to bottom * Set (None) option for columns to value of "" so properly not-set * Show island for prevailing metric, make it non-editable but mirrors main comparison * Removed Period & Aggregation on metrics (for now)
QA: You should now be able to create and save goals with metrics data
partially or fully filled-in. The UI in the goal editor is reorganized a
bit: Start Date is now part of the top section, and defaults to today if
not yet filled in. Related Datasets & Additional Details are moved below
the Metrics section. In the first metric, it will now have a center
comparison block, and the reduce/increase should not be directly
selectable, but copy what the top reduce/increase select says. You
should now be able to choose a date column for the Current & Baseline
sections (in addition to the value columns).

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>