FEATURE: GovStat: New header design * Big manage nav icons on left, with color * Header links go in dropdown menu * Remove previously-added icons in standard header links
QA: govstat.test has a new header with links to all the Manage pages as
icons in the top-left. The manage icons show a color underline on hover,
and the full icon and underline color when on that page. Standard header
links are in a drop-down menu in the top right.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>