FEATURE: GovStat: Manage reports styling * Refactored dashboard box styles to share with manage pages * Changed reports HTML to work with new styles * Show all reports here * Set common button style for govstat
QA: govstat.test: The Manage Reports page should look more like the
dashboard with large boxes for each report. It also shows all reports,
and no longer has a link to the admin page. Clicking a box allows you to
edit that report. Creating a new report is the last box in the list, in
action blue.

To-be-
Reviewed-by: Clint Tseng <clint.tseng@socrata.com>