FEATURE 8914: GovStat: Goal leaf page * Full rendering/styling of goal leaf page * Added option for custom javascript to generic_page * Changed progress_indicator globally to match latest design * Refactored govstat colors & header styles in SASS to use multiple places * Fixed line height in Additional Details in goals editor * When doing deepGet, auto-parse JSON strings * Multiple regexes can now be specified in a single string substitution * Regexes now properly handle \/ * Fixed string formatting 'u' in Rails * Related Datasets -> Related Visualizations in goals editor * Properly handle goal description in JS * Fixed date parsing in JS string subst
QA: On govstat.test, viewing a goal page should be nicely styled now.
You should be able to add Related Visualizations in the goals editor,
and they will appear on the goal page. Additional Details (description)
will appear below the charts on that page.
Also a general regression of canvas.test to make sure there aren't any
funny-looking strings

To-be-
Reviewed-by: Clint Tseng <clint.tseng@socrata.com>