BUG 9082: GovStat: Header fixups * Manage actions are hidden unless logged-in * When not logged-in, use user icon for menu * Make govstat header show up in printing (without actions) * Also fixed a problem with Button component sticking html in as text
QA: govstat.test: When not logged in, you shouldn't see the buttons in
the header for all the manage actions. The top-right icon will also show
a user (but still be a menu). The listed bug (code in the button) should
be fixed. The print preview for pages should show the basic header (dark
bar with the logo).

To-be-
Reviewed-by: Clint Tseng <clint.tseng@socrata.com>