BUG 9130, 8862: GovStat: Improve report creation * Path/URL is no longer required * It will auto-generate an appropriate path based on the title * Add the correct error message if title is missing * Make sure all paths start with / * In GovStat, hide the URL option
QA: govstat.test: Creating a report should not allow you to enter a path
anymore. It should also work even if you create two reports with the
same title.
canvas.test: When creating a DataSlate page, you can still manually
enter a path, but it is not required. Also if you enter a path that does
not start with /, it should automatically add that for you.
On both domains, 8862 should be fixed (correct error message if you
leave title blank).

To-be-
Reviewed-by: Clint Tseng <clint.tseng@socrata.com>