BUG 8245: DataSlate: Add currentUser to env * Only for privateData pages; else null
QA: canvas.test: The homepage should have the status line at the bottom
and includes 'User: none' whether or not you are logged-in. The "Private
Data" page should have a similar status line at top that has "User:"
with a different 4x4 id depending on which user you are logged-in as.

To-be-
Reviewed-by: Clint Tseng <clint.tseng@socrata.com>