BUG 9454: Fix maplayer filtering.
Okay, there's sort of a rabbit hole that we went down here, and it
affects 99.99% maplayers only, but not entirely.

What would happen is that the underlying layer dataset would get loaded
and the unified filter would get bound against it pretty immediately.
But because of the way maplayer datasets work with their child layer
datasets, at this point in time the child layer dataset doesn't actually
have any filters associated with it, so the filter interface loads
blank.

A couple of seconds later, the parent maplayer dataset finally decides
it's ready to load, at which point it tells the underlying layer dataset
what its filter should be. Unfortunately, the filter interface was never
designed to handle this case so it wouldn't do anything with that new
information.

CHANGE #1: add query_change listener to unified filter on all filtered
datasets; rerender filter interface when the query changes.
* There is an unfortunate side effect here: when you zoom or pan around,
  because of the way the filter keeps track of its state, some of the
  checkbox items may shift around. They should remain accurate, though,
  and the shifting isn't something I feel like I can fix.

Now we have a new problem. The maplayer datasets aren't terribly
respectful of the rest of the world, and also exhibit some bad patterns
that should never be employed. Specifically, they were caching, and deep
-merging queries. Queries, being array-indexed, should never ever be
deep-merged.  They should also never be cached away. What this meant was
that say we have a filter:

FilterCondition
  And
    0
      Equals
        0 column1
        1 value1

Now the maplayer has cached that away. Now we toggle a checkbox in the
filter interface:

FilterCondition
  And
    0
      Equals
        0 column1
        1 value2
    1
      Equals
        0 column1
        1 value1

Because of where suggested values get rendered, they go above existing
values, so the new value is at array index 0. Now we pan the map, and
the maplayer dataset goes "GREAT, let's deep-merge my cached, mutated
filter and the new filter you want!"

FilterCondition
  And
    0
      Equals
        0 column1
        1 value1
    1
      Equals
        0 column1
        1 value1

Oops. Now we're just filtering on value1 twice. That's not what we
wanted.

So, I made an assumption.

CHANGE #2: maplayer datasets should never internally care about any
filterConditions except for:
1. Initial assignment to child layer datasets.
2. the viewport namedFilter
So, I ripped out all the filter code that touched anything other than
viewport. That seems to have made the filters work okay.

CHANGE #3: now we seem to have filtering generally working but every
time you pan or zoom the filter interface reverts to simplified mode.
This is a long-standing behavioral quirk where despite the default state
of the filter interface being advanced, it never bothers to save that
fact away. This is why when you save a filter you've created the new
view defaults to being simplified. We should have fixed this long ago,
but it seemed equally justifiable as feature or bug, so we didn't. Now
we actually save off the advanced state of the filter, so there may be
some cases where you open an extant view that was simplified and it's
now magically advanced. Thus is life. Not a major enough behavioral
change that I would consider it something to address.

QA: See above. Ask me if anything doesn't make sense.

Reviewed-by: jeff.scherpelz@socrata.com