FEATURE: Listview for GovStat. (eg Hawaii-dash)
for the expanding section, it tries to show whichever of the following
first exist:
1. the first related visualization
2. the goal description
3. a description of the prevailing metric

QA: If you go to
https://animatedaardvarks.test-socrata.com/internal/orgs/2/domains/animatedaardvarks.test-socrata.com/site_config/64

and change the dashboard_layout key to read list instead of grid_flow
you should get a different looking layout on the homepage and /manage
pages of govstat sites. change back to grid_flow to test the current
view again. make sure it looks good in all browsers.

there are known issues with the related visualization chart rendering
weird. that will have to be something someone else looks at.

To-be-
Reviewed-by: jeff.scherpelz@socrata.com